put quick reply modal content inside HOC portal which fixes cursor jumping issue while typing
