+++ src/components/quickReplyModal/quickReplyModalStyles.ts
@@ -1,4 +1,6 @@
+import { Platform } from 'react-native';
 import EStyleSheet from 'react-native-extended-stylesheet';
+import { getBottomSpace } from 'react-native-iphone-x-helper';
 
 export default EStyleSheet.create({
   sheetContent: {
@@ -9,37 +11,36 @@ export default EStyleSheet.create({
     right: 0,
     zIndex: 999,
   },
-  container:{
-    flex:1,
+  container: {
+    flex: 1,
   },
   modalContainer: {
-    paddingBottom: 12,
+    paddingVertical: 4,
+    paddingBottom: Platform.OS === 'ios' ? getBottomSpace() : 12,
   },
-
 
   cancelButton: {
     marginRight: 10,
   },
   modalHeader: {
     paddingHorizontal: 12,
-    paddingVertical:8,
+    paddingVertical: 8,
   },
   titleBtnTxt: {
     fontSize: 18,
     fontWeight: 'bold',
     color: '$primaryBlack',
   },
-  summaryStyle:{
-    fontSize:16,
-    paddingHorizontal:16,
+  summaryStyle: {
+    fontSize: 16,
+    paddingHorizontal: 16,
     color: '$primaryDarkGray',
     fontWeight: '500',
   },
   inputContainer: {
     paddingHorizontal: 16,
-    paddingVertical:8,
-    height:120,
-
+    paddingVertical: 8,
+    height: 120,
   },
   textInput: {
     color: '$primaryBlack',
@@ -51,7 +52,7 @@ export default EStyleSheet.create({
     flexDirection: 'row',
     justifyContent: 'space-between',
     alignItems: 'center',
-    paddingHorizontal:16,
+    paddingHorizontal: 16,
   },
   commentBtn: {
     width: 100,
@@ -67,7 +68,7 @@ export default EStyleSheet.create({
     paddingTop: 10,
     paddingBottom: 0,
   },
-  
+
   accountTile: {
     height: 60,
     flexDirection: 'row',
@@ -78,9 +79,9 @@ export default EStyleSheet.create({
   avatarAndNameContainer: {
     flexDirection: 'row',
     alignItems: 'center',
-    paddingHorizontal:16,
-    paddingTop:12,
-    paddingBottom:8,
+    paddingHorizontal: 16,
+    paddingTop: 12,
+    paddingBottom: 8,
   },
   nameContainer: {
     marginLeft: 2,
